backend/requirements.txt
fastapi[all]
sqlalchemy
python-dotenv
bcrypt==4.0.1
passlib==1.7.4
python-jose[cryptography]
# alembic # Opcional para migraciones
uvicorn